backend/requirements.txt
django>=5.2.7
djangorestframework
django-cors-headers
mysqlclient